backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
qdrant-client
openai
python-dotenv
pydantic-settings
pydantic>=2.0
greenlet
requests
beautifulsoup4
lxml
